backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
channels>=4.0
channels-redis
psycopg2-binary
python-dotenv
gunicorn
django-cors-headers
daphne
dj-database-url
whitenoise[brotli]